requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.66.0
h11==0.12.0
importlib-metadata==4.6.1
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0

fastapi==0.78.0 
uvicorn==0.17.6 
pytest==7.1.1 
 
black==22.3.0 
flake8==4.0.1 
isort==5.10.1 
mypy==0.961
Pyrebase4==4.5.0